+++ common/SetupMessage.cpp
@@ -41,7 +41,7 @@ SetupMessage::DataMap parseGameData(const Value& data)
 
 } // End anonymous namespace
 
-SetupMessage::SetupMessage(uint16_t id, GameType gType, int pCount, int time, int score, DataMap&& data) :
+SetupMessage::SetupMessage(uint16_t id, GameType gType, int8_t pCount, int16_t time, int16_t score, DataMap&& data) :
 	Message(id),
 	gameType(gType),
 	playerCount(pCount),
@@ -79,8 +79,9 @@ std::unique_ptr<SetupMessage> SetupMessage::fromJSON(const Json::Value& object)
 
 	return std::unique_ptr<SetupMessage>(
 		new SetupMessage(msg->id,
-		                 (GameType)gameTypeValue.asInt(), playerCountValue.asInt(),
-		                 gameLengthValue.asInt(), winningScoreValue.asInt(), parseGameData(gameDataValue)));
+		                 (GameType)gameTypeValue.asInt(), (int8_t)playerCountValue.asInt(),
+		                 (int16_t)gameLengthValue.asInt(), (int16_t)winningScoreValue.asInt(),
+		                 parseGameData(gameDataValue)));
 }
 
 Json::Value SetupMessage::toJSON() const
